+++ KVstore/IVArray/IVEntry.h
@@ -28,8 +28,8 @@ public:
 	IVEntry();
 	
 	void setBstr(const Bstr* _value);
-	bool getBstr(char *& _str, unsigned long& _len, bool if_copy = true) const;
-	void setBstr(char *_str, unsigned long _len);
+	bool getBstr(char *& _str, unsigned& _len, bool if_copy = true) const;
+	void setBstr(char *_str, unsigned _len);
 	
 	void setStore(unsigned _store);
 	unsigned getStore() const;
